Revert "Add go dir and move sample"
