+++ .gitignore
@@ -1,3 +1,3 @@
 .idea/*
-*/.idea/
-*/*/.idea/
+helm/memcached-operator/.idea/*
+ansible/memcached-operator/.idea/*
