+++ memcached-operator/.gitignore
