+++ memcached-operator/deploy/crds/cache.example.com_memcacheds_crd.yaml
