+++ memcached-operator/deploy/crds/cache.example.com_v1alpha1_memcached_cr.yaml
