+++ memcached-operator/deploy/olm-catalog/memcached-operator/0.10.0/memcached-operator.v0.10.0.clusterserviceversion.yaml
