+++ memcached-operator/deploy/olm-catalog/memcached-operator/0.11.0/memcached-operator.v0.11.0.clusterserviceversion.yaml
