+++ memcached-operator/deploy/olm-catalog/memcached-operator/0.12.0/memcached-operator.v0.12.0.clusterserviceversion.yaml
