+++ memcached-operator/deploy/olm-catalog/memcached-operator/0.6.0/memcached-operator.v0.6.0.clusterserviceversion.yaml
