+++ memcached-operator/deploy/olm-catalog/memcached-operator/memcached-operator.package.yaml
