+++ memcached-operator/deploy/operator.yaml
