+++ memcached-operator/deploy/role.yaml
