+++ memcached-operator/pkg/apis/addtoscheme_cache_v1alpha1.go
@@ -1,7 +1,7 @@
 package apis
 
 import (
-	"github.com/operator-framework/operator-sdk-samples/go/memcached-operator/pkg/apis/cache/v1alpha1"
+	"github.com/operator-framework/operator-sdk-samples/memcached-operator/pkg/apis/cache/v1alpha1"
 )
 
 func init() {
