+++ memcached-operator/pkg/apis/cache/v1alpha1/memcached_types.go
