+++ memcached-operator/pkg/apis/cache/v1alpha1/register.go
