+++ memcached-operator/pkg/controller/memcached/memcached_controller_test.go
@@ -7,7 +7,7 @@ import (
 	"strconv"
 	"testing"
 
-	cachev1alpha1 "github.com/operator-framework/operator-sdk-samples/go/memcached-operator/pkg/apis/cache/v1alpha1"
+	cachev1alpha1 "github.com/operator-framework/operator-sdk-samples/memcached-operator/pkg/apis/cache/v1alpha1"
 
 	appsv1 "k8s.io/api/apps/v1"
 	corev1 "k8s.io/api/core/v1"
